src/main/java/com/eno/stationlang/service/compiler/error/ErrorListener.java
package com.eno.stationlang.service.compiler.error;

import java.util.BitSet;
import java.util.List;
import org.antlr.v4.runtime.ANTLRErrorListener;
import org.antlr.v4.runtime.Parser;
import org.antlr.v4.runtime.RecognitionException;
import org.antlr.v4.runtime.Recognizer;
import org.antlr.v4.runtime.atn.ATNConfigSet;
import org.antlr.v4.runtime.dfa.DFA;

public class ErrorListener implements ANTLRErrorListener {

  private final List<CompilationError> errors;

  public ErrorListener(List<CompilationError> errors) {
    this.errors = errors;
  }

  @Override
  public void syntaxError(
      Recognizer<?, ?> recognizer,
      Object offendingSymbol,
      int line,
      int charPositionInLine,
      String msg,
      RecognitionException e) {
    errors.add(new CompilationError(line, charPositionInLine, ErrorType.SYNTAX, msg));
  }

  @Override
  public void reportAmbiguity(
      Parser recognizer,
      DFA dfa,
      int startIndex,
      int stopIndex,
      boolean exact,
      BitSet ambigAlts,
      ATNConfigSet configs) {
    errors.add(new CompilationError(startIndex, stopIndex, ErrorType.AMBIGUITY, null));
  }

  @Override
  public void reportAttemptingFullContext(
      Parser recognizer,
      DFA dfa,
      int startIndex,
      int stopIndex,
      BitSet conflictingAlts,
      ATNConfigSet configs) {
    errors.add(new CompilationError(startIndex, stopIndex, ErrorType.CONTEXT, null));
  }

  @Override
  public void reportContextSensitivity(
      Parser recognizer,
      DFA dfa,
      int startIndex,
      int stopIndex,
      int prediction,
      ATNConfigSet configs) {
    errors.add(new CompilationError(startIndex, stopIndex, ErrorType.CONTEXT, null));
  }
}
